+++ babel.config.json
@@ -3,10 +3,8 @@
     [
       "@babel/preset-env",
       {
-        "targets": {
-          "node": true
-        }
-      }
+        "modules": false
+     }
     ],
     "@babel/preset-react",
     "@babel/preset-typescript"
